Creación del componente BannerModal.tsx
